requirements/production.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/production.in
#
appnope==0.1.0            # via ipython
asgiref==3.2.10           # via django
backcall==0.2.0           # via ipython
boto3==1.14.30            # via -r requirements/production.in
botocore==1.17.30         # via boto3, s3transfer
cached-property==1.5.1    # via django-stuff
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
decorator==4.4.2          # via ipython, traitlets
dj-database-url==0.5.0    # via -r requirements/production.in
dj-static==0.0.6          # via -r requirements/production.in
django-cors-headers==3.4.0  # via -r requirements/production.in
django-filter==2.3.0      # via -r requirements/production.in
django-stuff==0.5.1       # via -r requirements/production.in
django-widget-tweaks==1.4.8  # via -r requirements/production.in
django==3.0.8             # via -r requirements/production.in, django-cors-headers, django-filter, django-stuff, djangorestframework, rest-condition
djangorestframework==3.11.0  # via -r requirements/production.in, django-stuff, rest-condition
docutils==0.15.2          # via botocore
idna==2.10                # via requests
ipdb==0.13.3              # via -r requirements/production.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.16.1           # via ipdb
jedi==0.17.2              # via ipython
jmespath==0.10.0          # via boto3, botocore
numpy==1.19.1             # via -r requirements/production.in
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prettyconf==2.1.0         # via -r requirements/production.in
prompt-toolkit==3.0.5     # via ipython
psycopg2-binary==2.8.5    # via -r requirements/production.in
ptyprocess==0.6.0         # via pexpect
pygments==2.6.1           # via ipython
python-dateutil==2.8.1    # via botocore
pytz==2020.1              # via django
requests==2.24.0          # via -r requirements/production.in
rest-condition==1.0.3     # via -r requirements/production.in
s3transfer==0.3.3         # via boto3
six==1.15.0               # via python-dateutil, traitlets
sqlparse==0.3.1           # via django
static3==0.7.0            # via dj-static
traitlets==4.3.3          # via ipython
unipath==1.1              # via -r requirements/production.in, django-stuff
urllib3==1.25.10          # via botocore, requests
uwsgi==2.0.19.1           # via -r requirements/production.in
wcwidth==0.2.5            # via prompt-toolkit
whitenoise==5.1.0         # via -r requirements/production.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
